Issue/35 fixed bugs that resulted in failed tests and program crash (#39)
* fix/#35 , fixed a bug that resulted in that in test failing. Added an
extra function that tries to delete all folder in a map.

fixed a bug, resulting in program crash.
If the build_history folder contained a file with wrong formate the
program crashed. Thee program now ignores that file insteed.

fixed a bug. If the utilitytest was preformed before history
test it would fail. This was becuase the map used by utilitytest did not
exist. Utilitytest now creates the map if it does not exist.