feat/#16 Add request handlers for "/", "webhook", and "commit/?id=<commit hash id>"
ContinousIntegration class was modified to have three inner subclasses
of GeneralRequestHandler (which is a subclass AbstractRequestHandler).
Each subclass is created to handle requests of different types.

* WebhookRequestHandler handles request with path "/webhook/", which is used by
POST requests of github webhooks.
* CommitHashRequestHandler handles get requests with path "/commit?id=<commit hash id>",
where <commit hash id> is replaced with a commit hash id.
* ListHistoryRequestHandler handles get requests with path "/",
which is used to list the push history.

`//TODO` rows were added in the places of the methods that need to edited
to call methods that general the actual HTML responses.